requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.86.0
openai==2.30.0
faiss-cpu==1.13.2
rank-bm25==0.2.2
numpy==2.4.4
python-multipart==0.0.12
python-dotenv==1.0.1